(chore) update template to use u64 as the default answer

+++ template/src/main.rs
@@ -2,7 +2,7 @@ use std::fs;
 
 #[derive(Debug, PartialEq)]
 struct Answer {
-    answer: usize
+    answer: u64
 }
 
 trait InputGetter {
